Threading Bugfix: Call the login() method asynchronously.
The login() method is one of the only methods in the
rightscale.RightScale object that is pre-defined -- the rest of the
methods and attributes are dynamically created by blocking web calls.

We need to call the login() method asynchronously before we make any
other calls so that we pre-populate the connection objects properties.