Create a new Threadpool for every thread_coroutine() call.
The threadpool objects are supposed to have .shutdown() called when
they're done being used. During the Tornado unit tests we seem to run
into IOErrors on the threadpool objects. This patch creates a new fresh
threadpool object and uses it in a 'with' context, which should force it
to be shut down properly before the method ever returns.