Add 'warn_on_failure' option to BaseActor.
This new 'warn_on_failure' option changes the behavior of the BaseActor and
allow for actors to fail, but Kingpin scripts to continue to operate. It
works by returning True regardless of the execution status of the Actors
`_execute()` method.

The option defaults to False (same as previous behavior), and is sanity
checked by the JSON schema file.