Issue #114: Track execution time and print it out in debug statements.
This is just the beginning of this work.. but its a simple wrapper for
now to track the execution time of every actor and log it to a debug log
statement. In the future we can add optional printing of these
statements, global statistic tracking, and even some kind of reporting
at the end with the stats.