Use 'count=X' in rightscale.server_array.Launch actor when launching multiple instances.
This resolves issue #203 -- it uses a single API call when launching
multiple instances in RightScale. This leverages RightScales new API
feature (count=X) that allows RightScale to handle calling the "launch"
button over and over again on their own.

It doesn't save that much time, since RightScale still has to launch the
servers synchronously... but it reduces the number of API calls we make
and the room for failure.