Remove all calls to .login() in the RightScale actors.
Calling the python-rightscale.RightScale.login() method was ultimately a
waste of function calls in Python. The object automatically logs in when
its methods are executed, and after its logged in once, it doesn't log
in again.

Now that we're using a single RightScale object for all of our API
calls, this login happens far fewer times by letting it happen
automatically.

It is still possible that the object will make several API calls to
RightScale to login. This can happen if multiple Tornado Coroutines
fire off (nearly) asynchronously that all rely on the RightScale object
being logged in. Each one of these will trigger the same object to call
its .login() method. However, once a single method has finished and the
object is logged in, all future calls will use the existing login
credentials.