Allow setting of the 'strictness' of the rightscale.server_array.Clone actor.
The only reason that a "relaunch" of an existing ServerArray couldn't be
automated with existing actors was that the
rightscale.server_array.Clone actor was extremely strict about whether
the source/destination arrays already existed or not. By allowing this
to be configurable, you can write a script that acts on not-yet-existing
arrays.

Included in this change is an example of a "relaunch" Kingpin script.