Force Tornado < 4.1 due to a behavioral change in Exception logging.
Something in the way our code is calling multiple tasks and yielding on
them is causing Tornado 4.1 to log out all the exceptions again,
regardless of whether we're doing it or not. Until this is resolved,
pinning our Tornado to < 4.1.

I have opened an issue with the Tornado team asking for guidance:
https://github.com/tornadoweb/tornado/issues/1378