Prevent race-condition logging by quieting down ActorTimedOut exceptions.
When nested actors (an outer group actor, and an inner actor of some
kind) both throw an ActorTimedOut exception, the exception is invariably
raised from the outer actor. This causes the inner actor to activate a
new behavior in Tornado4 where that exception will be logged to the user
by Tornado itself. This behavior is normally good ... but not in this
case.

In this case, we just want to log out the outer-most ActorTimedOut
exception and let Kingpin do it. To fix this, we pass in ActorTimedOut
to the quiet_exceptions argument in Tornado.gen.with_timeout(), which
silences the duplicate alert.