Clean up all AWS Actors: Use retrying.retry with common backoff settings.
The AWS API is pretty sensitive when it comes to hammering it with
requests. We had all kinds of different retry logic in our code here as
we worked to deal with that.

This change wipes out all of that logic and puts it into a single
location -- the AWSBaseActor.thread() method. This provides us a single
common location to handle retrying. With this, we no longer use the
kingpin.actors.support.api._retry() method (which was sort of abused
here anyways).

The retrying logic is described in the kingpin.actors.aws.settings
module ... but has significant jitter, randomness, and exponential
backoffs applied.