Handle RightScales concept of 'array values'.
Per
http://reference.rightscale.com/api1.5/resources/ResourceServerArrays.html#create_parameters,
and
http://support.rightscale.com/12-Guides/RightScale_API_1.5/1_Overview/Parameters/index.html:

RightScale has a concept of input 'arrays' that are passed in as funny
looking key=value pairs. For example, passing in a schedule for
AutoScaling an array in RightScale looks like this:

  server_array[elasticity_params][schedule][][min_count]=1
  server_array[elasticity_params][schedule][][day]=Sunday
  server_array[elasticity_params][schedule][][max_count]=2
  server_array[elasticity_params][schedule][][time]=0700

  server_array[elasticity_params][schedule][][max_count]=2
  server_array[elasticity_params][schedule][][day]=Monday
  server_array[elasticity_params][schedule][][min_count]=1
  server_array[elasticity_params][schedule][][time]=0700

It should be noted that order only partially matters here. As long as
each group of 'day', 'max_count', 'min_count' and 'time' values are put
together, the specific order they are in does not matter. Note the
difference between the first 4 items and the second 4 items in the list
above.

This is extremely un-intuitive, but appears to be the way RightScale
works.