Bugfix: PackageCloud _delete() method broke out of loop too early.
The _delete() method had a nested raise gen.Return() in it that would be
executed as soon as we hit a minimum number of package versions for a
given package -- but that completely broke out of the loop, bypassing
all other package purging. I think this was left over cruft from some of
the refactoring.

Using 'break' seems appropriate here -- causing the 'for package in
package_versions' to break out, but keep Python executing the 'for name
in packages_list_to_delete' loop.