Bugfix: Need to use reverse=False to get the package lists correctly.
This was a nasty little gotcha -- unintuitively, you want to use
reverse=False to get the package lists in reverse chronological order.
Without doing this, the logic around which packages to keep and delete
is totally broken.

While debugging this, I added in a bunch of logging and comments to help
clarify things for users.