Allow Actor option-verification to easily validate against a list.
This change allows actors to define a simple class thats a list of valid
strings. When the option inputs are checked, if the supplied string
doesn't match the list in the class, then an error is raised.

Initial use case is for actors to have a 'STATE' requirement --
'present' or 'absent'.