Add group management support to the aws.iam.User actor.
The User, Group, Role and InstanceProfile actors all have some
differences when it comes to the more advanced properties they have. For
example. Users can be members of Groups, InstanceProfiles can be
assigned Roles.

This commit separates out the _execute() methods for each class so that
they can behave differently, and adds group membership management to the
User actor.