Full management of the ServerTemplate Runnable Bindings.
This code isn't as nice as I'd like. RightScale doesn't let us easily
add and control the order of scripts without purging all the previous
scripts. Basically any change to any of the scripts, revisions or
positions means that we need to purge all of the scripts and re-add
them.