Launch max_count machines if no `count` value specified. (#415)
This only affects the default launch count - in cases where count is not specified. Launching min value was originally chosen arbitrarily as "let's just launch the minimal needed servers." In practice if those servers are load bearing then the min value is only useful during least-traffic.

In the actual DSL file we've always specified count: {MAX_COUNT}. This is done such that transitioning traffic to a freshly-booted array does not suddenly hit a min_count array.

However... there is a huge difference between launching a count of max_count vs letting kingpin launch with default. The default launches up to the needed value, rather than new machines of that count.