Integrate an API Call Queue for cf.describe_stacks.
This will serialize all the requests to the CloudFormation describe_stacks API to better protect against rate limits.

This is a pretty complicated implementation, and I don't like how it integrates with testing at the moment.
Each test creates a new ioloop, breaking the existing ApiCallQueues, as the process_queue callback will not be re-entered. Haven't figured out a way around this yet other than clearly the ApiCallQueues between tests.