Remove use of Boto2, use Boto3 only, and deprecate under-maintained/used actors (#515)
**Huge Update**

This change is big - and very backwards-incompatible. I am officially deprecating a bunch of actors that are not actively developed or tested. In particular, I am deprecating ones based on `boto2` that is no longer supported, and we are not using. I am converting the actively used actors to `boto3`, and updating most of the test cases to use the Boto Stubber rather than mocks.

**Breaking!!**

Yes this is a huge breaking change... but we've been stuck trying to maintain old nasty compatibility with Python2, Boto2 and other old libraries for years now. The fact is that this is mostly used internally, and we need to modernize this code to be able to continue to iterate on it in a reasonable way.