Fix YAML Anchor handling (#516)
After we switched to using a more modern version of the YAML library, we ran into a problem where YAML Merge Maps started to fail. Others have seen the same problem:

  * https://github.com/zerwes/hiyapyco/pull/14/files
  * https://github.com/ssato/python-anyconfig/issues/87
  * https://github.com/yaml/pyyaml/issues/64
  * https://github.com/saltstack/salt/issues/2092
  * https://github.com/zerwes/hiyapyco/issues/7

In our case, we need to patch the Amazon-provided `construct_mapping` method to run through the `flatten_mapping()` function.